Added in-game console. Fixed lost exit event. Support for multiple trace backends.

+++ Engine/gui/messagebox.cpp
@@ -72,7 +72,7 @@ void MessageBox::setMessage(const string& text){
     }
     //concatenate word to current line
     line += word + " ";
-    chars += word.size()+1;
+    chars += (unsigned)word.size()+1;
   }
   msg_.push_back(line);
 }
